requirements.txt
azure-functions
azure-storage-blob
azure-storage-file-datalake
polars
pytest
pre-commit
